Progress: Finally Unblocked!
Table state is super wonky it's not updating the way i expect it to. Table Initialization took me for-fucking-ever

I figured out the issue! I was using Array.fill() which populated my table with references to the same array, that's why I was having such a hard time. Thanks ChatGPT!